+++ log.go
@@ -1,9 +1,14 @@
 package emperror
 
-// LogHandler accepts an logger instance and logs an error.
+import (
+	"github.com/go-kit/kit/log/level"
+	"github.com/goph/emperror/internal"
+)
+
+// logHandler accepts a logger instance and logs an error.
 //
 // Compatible with most level-based loggers.
-type LogHandler struct {
+type logHandler struct {
 	l logger
 }
 
@@ -12,15 +17,18 @@ type logger interface {
 	Log(keyvals ...interface{}) error
 }
 
-// NewLogHandler returns a new LogHandler.
+// NewLogHandler returns a new logHandler.
 func NewLogHandler(l logger) Handler {
-	return &LogHandler{l}
+	return &logHandler{level.Error(l)}
 }
 
 // Handle takes care of an error by logging it.
-func (h *LogHandler) Handle(err error) {
-	h.l.Log(
-		"level", "error",
-		"msg", err.Error(),
-	)
+func (h *logHandler) Handle(err error) {
+	keyvals := []interface{}{"msg", err.Error()}
+
+	if cerr, ok := err.(internal.ContextualError); ok {
+		keyvals = append(keyvals, cerr.Context()...)
+	}
+
+	h.l.Log(keyvals...)
 }