+++ log_test.go
@@ -3,30 +3,46 @@ package emperror_test
 import (
 	"testing"
 
-	"fmt"
+	"bytes"
+	"errors"
 
+	"github.com/go-kit/kit/log"
 	"github.com/goph/emperror"
+	"github.com/goph/emperror/internal"
 )
 
-type testLogger struct {
-	last string
-}
-
-func (l *testLogger) Log(keyvals ...interface{}) error {
-	l.last = keyvals[3].(string)
-
-	return nil
-}
-
 func TestLogHandler_Handle(t *testing.T) {
-	logger := &testLogger{}
+	t.Parallel()
+	buf := &bytes.Buffer{}
+	logger := log.NewLogfmtLogger(buf)
 	handler := emperror.NewLogHandler(logger)
 
-	err := fmt.Errorf("internal error")
+	err := errors.New("internal error")
 
 	handler.Handle(err)
 
-	if got, want := logger.last, err.Error(); got != want {
-		t.Fatalf("expected to log a specific error, received: %v", got)
+	if want, have := "level=error msg=\"internal error\"\n", buf.String(); want != have {
+		t.Errorf("\nwant: %shave: %s", want, have)
+	}
+}
+
+func TestLogHandler_HandleContext(t *testing.T) {
+	t.Parallel()
+	buf := &bytes.Buffer{}
+	logger := log.NewLogfmtLogger(buf)
+	handler := emperror.NewLogHandler(logger)
+
+	err := internal.ErrorWithContext{
+		Msg: "internal error",
+		Keyvals: []interface{}{
+			"a", 123,
+			"previous", errors.New("previous error"),
+		},
+	}
+
+	handler.Handle(err)
+
+	if want, have := "level=error msg=\"internal error\" a=123 previous=\"previous error\"\n", buf.String(); want != have {
+		t.Errorf("\nwant: %shave: %s", want, have)
 	}
 }