+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log
 
 
+## Unreleased
+
+### Added
+
+- Contextual logging of errors
+
+### Changed
+
+- Make error level default in Log handler
+- Do not export `LogHandler`
+
+
 ## 0.2.2 - 2017-07-07
 
 ### Added